requirements.txt
asgiref==3.8.1
Django==5.0.2
django-cors-headers==4.3.1
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.10
gunicorn==21.2.0
inflection==0.5.1
mysqlclient==2.2.4
packaging==25.0
pillow==10.2.0
PyJWT==2.10.1
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
uritemplate==4.2.0
whitenoise==6.6.0
